% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0 - main.R
\name{g_largest_component}
\alias{g_largest_component}
\title{Get largest connected component from graph}
\usage{
g_largest_component(g, mode = "strong")
}
\arguments{
\item{g}{igraph garph}

\item{mode}{passed to igraph::components (whether edge direction is considered,defaults to 'strong' (alernative: 'weak'))}
}
\description{
Get largest connected component from graph
}
